create and send endpoints are stubbed, moving forward with other endpoints and hashing implementations to spend as little time in stubbed state as possible
